# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "O2geeks Headless CMS",
    "description": "\n**Backend API for the O2geeks Branding System**\n\nThis API powers the decoupled content management backend. It provides robust RESTful endpoints for delivering and managing multi-format content.\n\n### Key Features:\n- **Authentication**: JWT-based secure access with Refresh Tokens.\n- **RBAC**: Strict Role-Based Access Control (Super Admin, Admin, Editor, User).\n- **Rate Limiting**: Distributed Redis-backed abuse protection.\n- **Media**: Dual-provider (Local + AWS S3) Multipart/Form-Data uploads.\n- **Engagement**: Zero-N+1 likes, comments, and favorites.\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/blogs": {
      "get": {
        "tags": [
          "Blogs"
        ],
        "summary": "List Blogs",
        "operationId": "list_blogs_api_v1_blogs_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Per Page"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_BlogOut_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Blogs"
        ],
        "summary": "Create Blog",
        "operationId": "create_blog_api_v1_blogs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_blog_api_v1_blogs_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blogs/{slug}": {
      "get": {
        "tags": [
          "Blogs"
        ],
        "summary": "Get Blog",
        "operationId": "get_blog_api_v1_blogs__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/blogs/{blog_id}": {
      "put": {
        "tags": [
          "Blogs"
        ],
        "summary": "Update Blog",
        "operationId": "update_blog_api_v1_blogs__blog_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Blog Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_blog_api_v1_blogs__blog_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Blogs"
        ],
        "summary": "Delete Blog",
        "operationId": "delete_blog_api_v1_blogs__blog_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "blog_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Blog Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/webhooks/{webhook_id}/logs": {
      "get": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Get Webhook Logs",
        "description": "View delivery logs for a specific webhook.",
        "operationId": "get_webhook_logs_api_v1_webhooks__webhook_id__logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Webhook Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_WebhookLogOut_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g., Validation Error or Duplicate Data)"
          },
          "401": {
            "description": "Unauthorized (Missing or Invalid JWT Token)"
          },
          "403": {
            "description": "Forbidden (Insufficient Permissions or Deactivated Account)"
          },
          "404": {
            "description": "Not Found (The requested resource does not exist)"
          },
          "429": {
            "description": "Too Many Requests (Rate limit exceeded for this endpoint)"
          },
          "500": {
            "description": "Internal Server Error (Unexpected system failure)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlogOut": {
        "properties": {
          "likes_count": {
            "type": "integer",
            "title": "Likes Count",
            "default": 0
          },
          "comments_count": {
            "type": "integer",
            "title": "Comments Count",
            "default": 0
          },
          "is_liked": {
            "type": "boolean",
            "title": "Is Liked",
            "default": false
          },
          "status_changed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Changed At"
          },
          "status_changed_by_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Changed By Id"
          },
          "status_change_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Change Reason"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "cover_image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "author",
          "content",
          "excerpt",
          "cover_image",
          "category",
          "tags",
          "status",
          "published_at",
          "created_at",
          "updated_at"
        ],
        "title": "BlogOut",
        "description": "Response model for a blog post."
      },
      "Body_create_blog_api_v1_blogs_post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "author": {
            "type": "string",
            "maxLength": 150,
            "title": "Author"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "JSON array string, e.g. [\"python\",\"fastapi\"]"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus",
            "default": "draft"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "cover_image": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image",
            "description": "Cover image"
          }
        },
        "type": "object",
        "required": [
          "title",
          "author",
          "content"
        ],
        "title": "Body_create_blog_api_v1_blogs_post"
      },
...
      "ContentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "in_review",
          "changes_requested",
          "approved",
          "scheduled",
          "published",
          "unpublished",
          "archived"
        ],
        "title": "ContentStatus",
        "description": "Publication status for all content types."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_BlogOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BlogOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[BlogOut]"
      },
...
      "PaginatedResponse_WebhookLogOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WebhookLogOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "per_page": {
            "type": "integer",
            "title": "Per Page"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "title": "PaginatedResponse[WebhookLogOut]"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}